#19 ixed up the service register logic to not check for .notRegistered which will always fail if the OS has never seen the domain before.
added `Program` key to plists. This should eventually become `BundleProgram` for full services management support but this would break backwards compatibility so we're leaving the full path in for now
services management also seems to have issue with agents destined for use at the login window. Until (if) this is fixed, we'll do a manual copy of that agent out of the app bundle into `/Library/LaunchAgents` to ensure it loads.